Handle singular mgmt-ip
If only one mgmt-ip is returned by lldpcli, the JSON output is formatted
in such a way that instead of an array with one element of type string,
it contains the string directly.

This trips up go's type system, which uses static types and therefore
needs extra code to handle this case.