Refactor lldpcli JSON handling (Part 1)
lldpcli JSON output has consistently proved to be rather insane, however
I've so far been using it as internal data structure for passing around
LLDP information internally.

The refactoring effort is supposed to clean it up by defining my own
internal data structure which is more sane, and use the lldpcli JSON
format only when getting the information out of the tool.

This commit adds an Lldpcli prefix to all lldpcli related datastructures
to be able to pick them out of the code more quickly in step 2.