Fix lldpcli JSON (chassis capability) on microsens
This switch only reports a single member in the capabilities array.
As usual, lldpcli serializes an array with a single member into
just that member not wrapped in an array, which confuses the JSON
parser. These capabilities are not used anywhere so I won't parse
them until we need them.