Move entire topologyd source into subdir topologyD
godoc refuses to generate source level documentation for the main
package "because it is a command, not a library".

For more information see https://github.com/golang/go/issues/5727

The package name "topologyd" is also unavailable because it's the
designated output file for "go build". To avoid the conflict, the
weird capitalization was chosen as an easy straightforward fix.

Therefore this commit renames main() to topologyd.Main() and creates
a new "main" package which only contains a main function that calls
topologyD.Main().

Go developers, sometimes your decision making process is really
very backwards...