Parallelize gather loops
gather_neighbors_from_nodes() and (*NodeMap).gather_stp_states() now run
in parallel.

On my test setup there are 6 machines:
4× topologyd (REST API)
1× managed switch (Microsens, SNMP)
1× extra machine refusing connections on both ports

The SNMP switch takes the majority of the query time. With the change,
the lookup on this small network takes 6 seconds rather than 7.

With more managed switches and other machines involved, this will
significantly decrease the wait time for the gathering process, e.g. if
there are two switches with slow SNMP the wait time will be 6 seconds
instead of 12.